+++ src/ImportedBuilding.h
@@ -35,14 +35,17 @@ public:
     static int noBottom;
 
     ImportedBuilding() = delete;
-    ImportedBuilding(std::unique_ptr<nlohmann::json>& buildingJson, Point3VectorPtr& importedBuildingPts, const int internalID);
+    ImportedBuilding(std::unique_ptr<nlohmann::json>& buildingJson,
+                     PointSet3Ptr& importedBuildingPts, const int internalID);
     ImportedBuilding(Mesh& mesh, const int internalID);
     ~ImportedBuilding();
 
-    virtual void  reconstruct() override;
-    virtual void  reconstruct_flat_terrain() override;
+    virtual double get_elevation() override;
+    virtual void   reconstruct() override;
+    virtual void   reconstruct_flat_terrain() override;
 
-    void append_nonground_part(const std::shared_ptr<ImportedBuilding>& other);
+    double get_height();
+    void   append_nonground_part(const std::shared_ptr<ImportedBuilding>& other);
 
     const nlohmann::json& get_building_json() const;
     const std::string&    get_parent_building_id() const;
@@ -54,14 +57,12 @@ public:
 
 protected:
     std::unique_ptr<nlohmann::json>  _buildingJson;
-    double                           _avgFootprintHeight;
     std::vector<int>                 _footprintIdxList;
     std::vector<std::vector<int>>    _footprintPtsIdxList;
     std::string                      _parentBuildingID;
     bool                             _appendToBuilding;
     bool                             _trueHeight;
     int                              _lodIdx;
-    Point3VectorPtr                  _dPts;
 
     void check_simplicity(Polygon_2& ring);
     void polyset_to_polygon(const CGAL::Polygon_set_2<CGAL::Epeck>& polySet);