+++ src/PointCloud.h
@@ -42,11 +42,10 @@ public:
     ~PointCloud();
 
     void random_thin_pts();
-    void create_flat_terrain(const PolyFeatures& lsFeatures);
+    void create_flat_terrain(const PolyFeaturesPtr& lsFeatures);
     void set_flat_terrain();
     void smooth_terrain();
-    void flatten_polygon_pts(const PolyFeatures& lsFeatures);
-    SearchTreePtr make_search_tree_buildings();
+    void flatten_polygon_pts(const PolyFeaturesPtr& lsFeatures);
     void read_point_clouds();
 
     Point_set_3& get_terrain();