+++ thirdparty/CGAL/include/CGAL/Search_traits_adapter.h
@@ -1,10 +1,11 @@
+
 // Copyright (c) 2011 GeometryFactory (France).
 // All rights reserved.
 //
 // This file is part of CGAL (www.cgal.org).
 //
-// $URL: https://github.com/CGAL/cgal/blob/v5.5/Spatial_searching/include/CGAL/Search_traits_adapter.h $
-// $Id: Search_traits_adapter.h 8166579 2021-10-11T19:58:07+02:00 Mael Rouxel-Labbé
+// $URL$
+// $Id$
 // SPDX-License-Identifier: GPL-3.0-or-later OR LicenseRef-Commercial
 //
 //
@@ -142,14 +143,14 @@ public:
     typedef typename boost::property_traits<PointPropertyMap>::value_type
     Point;
 
-    std::shared_ptr<Point> point;
-    std::size_t idx;
+    Point point;
+    std::size_t idx = 0;
 
   public:
 
-    No_lvalue_iterator() : point(NULL), idx(0) { }
-    No_lvalue_iterator(const Point& point) : point(new Point(point)), idx(0) { }
-    No_lvalue_iterator(const Point& point, int) : point(new Point(point)), idx(Base::Dimension::value) { }
+    No_lvalue_iterator() : point() { }
+    No_lvalue_iterator(const Point& point) : point(point) { }
+    No_lvalue_iterator(const Point& point, int) : point(point), idx(Base::Dimension::value) { }
 
   private:
 
@@ -157,18 +158,15 @@ public:
     void increment()
     {
       ++idx;
-      CGAL_assertion(point != std::shared_ptr<Point>());
     }
     void decrement()
     {
       --idx;
-      CGAL_assertion(point != std::shared_ptr<Point>());
     }
 
     void advance(std::ptrdiff_t n)
     {
       idx += n;
-      CGAL_assertion(point != std::shared_ptr<Point>());
     }
 
     std::ptrdiff_t distance_to(const No_lvalue_iterator& other) const
@@ -185,7 +183,7 @@ public:
     dereference() const
     {
       // Point::operator[] takes an int as parameter...
-      return const_cast<Dereference_type&>((*point)[static_cast<int>(idx)]);
+      return const_cast<Dereference_type&>(point[static_cast<int>(idx)]);
     }
 
   };